Bitcoin Classic Token  eth classic erc-223
Smart Contract
Typically, ERC223 tokens will grant all tokens to the owner or will have an ICO which
demands that amounts of ETC be sent to the owner for an initial offering of tokens.
Instead of granting tokens to the 'contract owner', all BCT tokens are locked within
the smart contract initially. These tokens are dispensed, 50 at a time, by calling the
function 'mint' and using Proof of Work, similar to mining bitcoin classic.